// app/src/model/sibiu-schools.ts
export const SIBIU_SCHOOLS = [
  {
    "code": "3261100049",
    "name": "ȘCOALA GIMNAZIALĂ NR. 2 SIBIU",
    "n": 20,
    "romana": 0.07506879999999999,
    "matematica": -0.3082816
  },
  {
    "code": "3261100058",
    "name": "ȘCOALA GIMNAZIALĂ NR. 4 SIBIU",
    "n": 63,
    "romana": 0.33914666666666665,
    "matematica": 0.459261935483871
  },
  {
    "code": "3261100117",
    "name": "ȘCOALA GIMNAZIALĂ \"REGINA MARIA\" SIBIU",
    "n": 97,
    "romana": 0.2550308661417322,
    "matematica": 0.6011407874015754
  },
  {
    "code": "3261100189",
    "name": "LICEUL TEORETIC \"CONSTANTIN NOICA\" SIBIU",
    "n": 59,
    "romana": -0.054158651685393154,
    "matematica": 0.10853168539325848
  },
  {
    "code": "3261100198",
    "name": "ȘCOALA GIMNAZIALĂ \"REGELE FERDINAND\" SIBIU",
    "n": 55,
    "romana": 0.11908894117647034,
    "matematica": -0.2724880000000001
  },
  {
    "code": "3261100239",
    "name": "ȘCOALA GIMNAZIALĂ NR. 18 SIBIU",
    "n": 84,
    "romana": -0.016291228070175304,
    "matematica": 0.42218175438596495
  },
  {
    "code": "3261101222",
    "name": "ȘCOALA GIMNAZIALĂ \"NICOLAE IORGA\" SIBIU",
    "n": 129,
    "romana": 0.4638966037735852,
    "matematica": 0.8121084276729558
  },
  {
    "code": "3261101412",
    "name": "ȘCOALA GIMNAZIALĂ NR. 1 SIBIU",
    "n": 55,
    "romana": -0.014080470588235335,
    "matematica": -0.18547905882352958
  },
  {
    "code": "3261101638",
    "name": "ȘCOALA GIMNAZIALĂ NR. 10 SIBIU",
    "n": 15,
    "romana": 0,
    "matematica": 0
  },
  {
    "code": "3261101656",
    "name": "LICEUL DE ARTĂ SIBIU",
    "n": 38,
    "romana": 0.1264817647058824,
    "matematica": -0.11470058823529394
  },
  {
    "code": "3261101742",
    "name": "CENTRUL ȘCOLAR DE EDUCAȚIE INCLUZIVĂ NR. 2 SIBIU",
    "n": 0,
    "romana": 0,
    "matematica": 0
  },
  {
    "code": "3261101751",
    "name": "ȘCOALA GIMNAZIALĂ \"RADU SELEJAN\" SIBIU",
    "n": 78,
    "romana": 0.28556000000000015,
    "matematica": 0.22632148148148148
  },
  {
    "code": "3261101801",
    "name": "ȘCOALA GIMNAZIALĂ NR. 25 SIBIU",
    "n": 45,
    "romana": 0.4660293333333336,
    "matematica": 0.4775706666666668
  },
  {
    "code": "3261101837",
    "name": "COLEGIUL NAȚIONAL \"GHEORGHE LAZĂR\" SIBIU",
    "n": 58,
    "romana": 0.3307845454545456,
    "matematica": 0.7851795454545448
  },
  {
    "code": "3261101968",
    "name": "COLEGIUL NAȚIONAL PEDAGOGIC \"ANDREI ȘAGUNA\" SIBIU",
    "n": 28,
    "romana": 0.15605517241379313,
    "matematica": 0.008972413793103394
  },
  {
    "code": "3261102087",
    "name": "CENTRUL ȘCOLAR DE EDUCAȚIE INCLUZIVĂ NR. 1 SIBIU",
    "n": 0,
    "romana": 0,
    "matematica": 0
  },
  {
    "code": "3261102101",
    "name": "ȘCOALA GIMNAZIALĂ \"IOAN SLAVICI\" SIBIU",
    "n": 52,
    "romana": 0.2750912195121953,
    "matematica": 0.6608385365853656
  },
  {
    "code": "3261102137",
    "name": "COLEGIUL NAȚIONAL \"OCTAVIAN GOGA\" SIBIU",
    "n": 111,
    "romana": 0.10940879432624137,
    "matematica": 0.34749390070921954
  },
  {
    "code": "3261102155",
    "name": "ȘCOALA GIMNAZIALĂ NR. 23 SIBIU",
    "n": 17,
    "romana": 0,
    "matematica": 0
  },
  {
    "code": "3261102336",
    "name": "ȘCOALA GIMNAZIALĂ NR. 13 SIBIU",
    "n": 74,
    "romana": 0.49544807692307674,
    "matematica": 0.8279507692307688
  },
  {
    "code": "3261102422",
    "name": "ȘCOALA GIMNAZIALĂ NR. 8 SIBIU",
    "n": 66,
    "romana": 0.20059333333333348,
    "matematica": 0.5126991666666667
  },
  {
    "code": "3261102517",
    "name": "ȘCOALA GIMNAZIALĂ NR. 21 SIBIU",
    "n": 73,
    "romana": 0.3484213592233012,
    "matematica": 0.3347557281553399
  },
  {
    "code": "3261103183",
    "name": "ȘCOALA GIMNAZIALĂ \"ION LUCA CARAGIALE\" SIBIU",
    "n": 53,
    "romana": 0.29717734939759055,
    "matematica": 0.489356144578313
  },
  {
    "code": "3261203873",
    "name": "LICEUL ROMÂNO - FINLANDEZ ERI SIBIU",
    "n": 15,
    "romana": 0,
    "matematica": 0
  },
  {
    "code": "3262100958",
    "name": "ȘCOALA GIMNAZIALĂ NR. 11 SIBIU",
    "n": 5,
    "romana": 0,
    "matematica": 0
  },
  {
    "code": "3262100967",
    "name": "ȘCOALA GIMNAZIALĂ NR. 12 SIBIU",
    "n": 12,
    "romana": 0,
    "matematica": 0
  }
] as const;